+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -21,6 +21,10 @@
 - Add new API `store::Store::room_keys_received_stream` to provide
   updates of room keys being received.
 
-- Add new method `identities::device::Device::first_time_seen_ts` 
+- Add new method `identities::device::Device::first_time_seen_ts`
   that allows to get a local timestamp of when the device was first seen by
   the sdk (in seconds since epoch).
+
+- When rejecting a key-verification request over to-device messages, send the
+  `m.key.verification.cancel` to the device that made the request, rather than
+  broadcasting to all devices.